requirements.txt
# Core dependencies
pandas>=2.2.2
numpy>=1.26.4
tenacity>=8.5.0
python-dotenv>=1.0.1
requests>=2.32.3

# FastAPI Backend
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0

# Streamlit Frontend
streamlit>=1.36.0
altair>=5.3.0

# Cloud (optional)
google-cloud-bigquery>=3.25.0
google-cloud-storage>=2.17.0
google-auth>=2.34.0

# LLM (optional)
openai>=1.45.0
Pillow>=10.4.0
# Optional OCR support (requires Tesseract installed separately)
pytesseract>=0.3.13

# Multi-platform ad integrations
google-ads>=22.0.0

# A/B testing statistical analysis
scipy>=1.10.0
